Update configuration for charge data to include charge type
