+++ src/BookKeeping/Msape.BookKeeping.Data/ChargeConfiguration.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -11,15 +10,5 @@ namespace Msape.BookKeeping.Data
         public TransactionType TransactionType { get; set; }
         public Currency Currency { get; set; }
         public List<ChargeData> Data { get; set; } = new List<ChargeData>();
-    }
-
-    public class ChargeData
-    {
-        public TransactionType ChargeType { get; set; }
-        public DateTime FromDate { get; set; }
-        public DateTime? ToDate { get; set; }
-        public decimal MinAmount { get; set; }
-        public decimal MaxAmount { get; set; }
-        public decimal ChargeAmount { get; set; }
     }
 }
